Disable the Sorbet department by default
Even though we have
```
AllCops:
  DisabledByDefault: true
```
The option is not respected for custom cops. This leads to Sorbet cops being ran on repos that are not using Sorbet on Policial,
because Policial loads the custom cops.
This kind of sucks for local usage because it prints a warning about unknown department, but we can remove this once upstream
fixes this issues.