Make the style guide a gem
This will make easier to use in the projects since change to this
repository will not automatically be applied to users and we don't need
to deal with caches anymore.

We can also define which minimal version of rubocop we support now.

As we upgrade our gems using dependabot by default, every time we
release a new version of the style guide, the projects will receive the
new version as a PR.

Closes #140.