Allow shadowing outer local variables
This lint was created to replace a warning that Ruby removed.

The warning existed because there was a change in behavior introduced in
version 1.9. Previously reusing a local variable resulted in the outer
variable taking the value of the inner variable. After that change,
there is shadowing, so they added a warning, which was finally removed
in 2.6: https://bugs.ruby-lang.org/issues/12490