Enable `Style/ExplicitBlockArgument` cop
The code is disabled by default because in some cases
it could change the block arity which might fail if the block is
actualy a lambda.

However I enabled it on Rails and shopify-core and it went just fine.

Since this cop is beneficial for production performance and
for triming CI backtraces a bit, I think we should enable it.