Stop disabling Style/MutableConstant
I couldn't find any discussion about this cop, but I believe it's a positive one.

First there's the correctness argument. If a constant is mutable is it really a constant?

Then there's a performance argument. Hashes unless frozen are mutated uppon iteration
(their internal `iter_level` attribute is incremented). This cause constant hashes that
are in a clean CoW region to invalidate the entire 4kiB memory page.