Set Style/TrailingCommaInArguments's EnforcedStyleForMultiline to comma
Differently from the Hash and Array versions of this rule, this one
uses the existence of parentheses to determine if a comma is needed
or not, not if the method is multiline. As our style guide require
parentheses for all method calls, so this rule can't use the
consistent_comma option.