Skip dumping config if required token is missing
We need a specific token to be able to make commits that will trigger
other actions. However, Dependabot commits use different secrets.

Therefore, this makes dumping the config conditional on having the
token. It is structured as a separate job due to limitations around the
availability of secrets in `if` contexts, and to avoid having the
conditional on every step.