Memoization strategy
Memoization is useful if you want to run a command once but be able to access that value many times.

However, the popular `||=` operator is not great for when your command might return a falsey value. By guarding memoization instance methods with `if defined?`, we make sure it only runs once.