Compare RuboCop version against strings
As of Rubygems 3.5.6 (included in Ruby 3.2 onwards, and often available
in environments with older Ruby versions where Rubygems has been
updated), `Gem::Version#<=>` supports passing in a `String`, instead of
needing to create a `Gem::Version` to pass in.

If we temporarily add a backport, we can use this approach, which makes
the version checks in the YAML+ERB config files more straightforward.

The backport is done in a separate file so we can leverage the
deduplication `require` provides out of the box, so as not to trigger a
method redefinition warning. It also means we'll be able to import it in
our upcoming plugin config files, rather than duplicating it.