Add missing files to `gemspec`
2.15.0 is broken because we're failing to include the backport patch
which we try to `require` at the top of `rubocop.yml`.

This updates the `gempspec` to include all ruby files under `lib`, as
well as the `README.md`. It also switches to use a wildcard to match the
`rubocop.yml` and `rubocop-cli.yml` files, which will be useful when we
add extension configs.