Add Lint/NoReturnInMemoization cop
This cop checks for the use of `return` with a value in `begin..end` blocks
in the context of instance variable assignment such as memoization. Using
`return` with a value in these blocks can lead to unexpected behavior as
the `return` will exit the method and not set the value of the instance
variable.